% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colength.R
\name{Colength}
\alias{Colength}
\title{Colength}
\usage{
Colength(a, b)
}
\arguments{
\item{a}{column 1, numeric assumed}

\item{b}{column 2, numeric assumed}
}
\value{
Print a statement telling you which numeric column has more N/A's
}
\description{
Colength
}
