% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Difference.R
\name{Difference}
\alias{Difference}
\title{Difference b/n Max and min}
\usage{
Difference(df, a)
}
\arguments{
\item{df}{dataframe with numeric valeus}

\item{a}{column we want to find the difference b/n max and min}
}
\value{
returns a number
}
\description{
Difference b/n Max and min
}
\examples{
If max and min are 3 and 1, it should return 2
}
