% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drop.R
\name{drop_columns}
\alias{drop_columns}
\title{Drops NA values}
\usage{
drop_columns(df, threshold)
}
\arguments{
\item{df}{This is a structured data frame with any data type valeus}

\item{threshold}{Cutofff for missing data}
}
\value{

}
\description{
Drops NA values
}
\examples{
Inputting some data frame and a threshold number
}
