% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputateNA.R
\name{NA_Imputate}
\alias{NA_Imputate}
\title{Imputate NA in a column with 0}
\usage{
NA_Imputate(df, column_name)
}
\arguments{
\item{df}{This is a dataframe that houses the columns/variables}

\item{column_name}{this is the name of the column and variable in question}
}
\value{
column but with the NA values now turned to 0
}
\description{
Imputate NA in a column with 0
}
\examples{
Inputting a dataframe and a column name
}
