% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumSquares.R
\name{SumSquares}
\alias{SumSquares}
\title{SumSquares}
\usage{
SumSquares(x, y)
}
\arguments{
\item{x}{an X value}

\item{y}{a Y value}
}
\value{
Returns sum of the squares of x and y
}
\description{
SumSquares
}
\examples{
X=1, Y=1, SumSquares = 2
}
